split spec generation to separate class
