+++ src/ingest/test/TestIngestion.cpp
@@ -21,6 +21,7 @@
 #include <pg_query.h>
 #include <storage/NFS.h>
 
+#include "execution/meta/SpecProvider.h"
 #include "ingest/IngestSpec.h"
 #include "ingest/SpecRepo.h"
 #include "meta/ClusterInfo.h"
@@ -31,6 +32,7 @@ namespace nebula {
 namespace ingest {
 namespace test {
 
+using nebula::execution::meta::genPatternSpec;
 using nebula::ingest::SpecRepo;
 using namespace nebula::meta;
 
@@ -107,12 +109,12 @@ TEST(IngestTest, TestTimePatternSpecGeneration) {
   sr.refresh(ci);
 
   // test macro parser
-  const auto& specs = ci.tables();
-  for (auto spec : specs) {
-    const auto type = spec->timeSpec.type;
-    const auto name = spec->name;
+  const auto& tables = ci.tables();
+  for (auto table : tables) {
+    const auto type = table->timeSpec.type;
+    const auto name = table->name;
     if (type == meta::TimeType::MACRO && name == "nebula.hourly") {
-      const auto sourceInfo = nebula::storage::parse(spec->location);
+      const auto sourceInfo = nebula::storage::parse(table->location);
       std::string macroStr = "hour";
 
       int pos = sourceInfo.path.find(macroStr);
@@ -121,21 +123,28 @@ TEST(IngestTest, TestTimePatternSpecGeneration) {
       EXPECT_EQ(pathTemplate.substr(pos, strlen("2020-01-01")), "2020-01-01");
 
       // scan four hour ago
-      auto ingestSpec = std::vector<std::shared_ptr<IngestSpec>>();
+      std::vector<SpecPtr> specs;
       // always include current hour
-      SpecRepo::genPatternSpec(meta::PatternMacro::HOURLY,
-                               sourceInfo.path,
-                               common::Evidence::HOUR_SECONDS * 4,
-                               "1.0", spec, ingestSpec);
+      genPatternSpec(meta::PatternMacro::HOURLY,
+                     sourceInfo.path,
+                     common::Evidence::HOUR_SECONDS * 4,
+                     "1.0",
+                     table,
+                     specs);
+
+      // no file to be listed from s3 with no-existing path
+      EXPECT_EQ(specs.size(), 0);
     } else if (type == meta::TimeType::MACRO && name == "nebula.daily") {
-      const auto sourceInfo = nebula::storage::parse(spec->location);
+      const auto sourceInfo = nebula::storage::parse(table->location);
       std::string macroStr = "date";
       // scan an hour ago
-      auto ingestSpec = std::vector<std::shared_ptr<IngestSpec>>();
+      auto specs = std::vector<SpecPtr>();
+
       // TODO (chenqin): mock s3 filesystem list
       // expect cover last day
-      SpecRepo::genPatternSpec(meta::PatternMacro::DAILY, sourceInfo.path, 3600,
-                               "1.0", spec, ingestSpec);
+      genPatternSpec(meta::PatternMacro::DAILY, sourceInfo.path, 3600, "2.0", table, specs);
+      // no file to be listed from s3 with no-existing path
+      EXPECT_EQ(specs.size(), 0);
     }
   }
 }
